requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
jupyterlab
torch
torchvision
torchaudio
transformers[torch]
Pillow
simple-pid
pandas
pywin32
mss
keyboard
opencv-python
matplotlib
seaborn
scipy